Commit 2
I have used these initial commits as a writing editor. Soon, I will get into the code and styling.

This has been incredibly helpful to understand how to create in code.

By taking a form I understand (writing) and can flow in, and then throw in a line break or an h2 tag has actually made me see html quite differently in the little time I've spent crafting this index.html page.

I have found writing a website so tough in the past, because in my mind the content and code were separate. This is the first stage of realizing they are the same.